+++ src/lang/libinterpreter/interpreter.cpp
@@ -532,8 +532,8 @@ std::any Interpreter::visit(AstNode::Expression::Access& ast) {
 std::any Interpreter::visit(AstNode::Expression::Literal& ast) {
     std::any literal;
     const auto convert = overloads {
-        [&literal](int64_t value) {literal = BlsType((int)value);},
-        [&literal](double value) {literal = BlsType((float)value);},
+        [&literal](int64_t value) {literal = BlsType((int64_t)value);},
+        [&literal](double value) {literal = BlsType((double)value);},
         [&literal](auto value) { literal = BlsType(value); }
     };
     std::visit(convert, ast.getLiteral());
@@ -599,11 +599,11 @@ std::any Interpreter::visit(AstNode::Specifier::Type& ast) {
             break;
 
             case TYPE::int_t:
-                return BlsType(int());
+                return BlsType(int64_t());
             break;
 
             case TYPE::float_t:
-                return BlsType(float());
+                return BlsType(double());
             break;
 
             case TYPE::string_t:
